small bug understimating count in kcount
the detection of substitution in degen was non working. two bugs were present.
First the loop start with foreach (0 .. ( 1- length($kmer2) )) with yield a negative number and does not loop.
-> changed to foreach (0 .. ( length($kmer2) -1 ))

Secondly, the string substitution was problematic 
my $degen_kmer = substr($kmer2, $_, 1, "[ACTGN]"); 
this line modify 2 variable
$degen_kmer is assigned to the excised base
and $kmer2 is modified in place
a solution is to make a copy of the kmer at each loop to modify it in place and use it in the regexp.

I tested this solution on a small subset and it seems to yield a coherent and consistent result.